Refactor module validation into its own module
